api_gateway/requirements.txt
fastapi
uvicorn
kafka-python-ng